+++ src/main/java/sql/insert/team_insert.java
@@ -1,6 +1,7 @@
-package sql;
+package sql.insert;
 
 import lombok.extern.slf4j.Slf4j;
+import sql.DataSource;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
